requirements.txt
# Core dependencies
asyncio-mqtt==0.16.1
aiohttp==3.9.1
python-json-logger==2.0.7

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AWS S3
boto3==1.34.10
aioboto3==12.3.0

# Document processing
PyPDF2==3.0.1
pandas==2.1.4
pdfplumber==0.10.3

# Validation
pydantic==2.5.3
jsonschema==4.20.0

# Environment
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

